Add build infrastructure for native Apple Silicon host
On every release tag and manual trigger when the "Include builds on non-free runners" checkbox is checked, make the
Arduino IDE build for native Apple Silicon host in addition to the builds that are always generated by the "Arduino IDE"
GitHub Actions workflow.

Previously, the build workflow only produced a build for x86-64 (AKA "Intel") macOS hosts. These can be used on Apple
Silicon machines via the Rosetta 2 translation software, but there is a performance impact. Native Apple Silicon builds
had to be produced and published manually.

GitHub gives unlimited use of the basic runner machines for workflow runs in public repositories. The macOS ARM
architecture is only provided in runner machines which are classified as "larger runner". Use of these runners is
charged on a per-minute basis ($0.16/minute), without any of the free allowances GitHub provided for the normal runners.
For this reason, native Apple Silicon builds must be generated only when there is compelling reason to do so. Such a
build is needed for every release, so the workflow is configured to always generate the builds when triggered by a tag.
In addition to releases, Apple Silicon tester builds for pull requests that might have special implications for this
target. For this reason, the workflow is configured to allow Apple Silicon builds to be triggered manually by a
repository maintainer.