Use standardized identifier for ARM target build filenames
A convention has been established of using "ARM" in the host target identifiers (e.g., "ARM64") of the build artifact
filenames in Arduino tooling projects. Previously the Arduino IDE 2.x project departed from that convention by instead
using "arm" (e.g., "arm64").

The inconsistency is hereby resolved by using "ARM" in the filenames of the artifacts generated by the build process as
well as in the names of the tester build workflow artifacts.