Add `PAID_RUNNER_BUILD_DATA` environment variable back to build workflow
The build workflow produces builds for a range of target host architectures, including macOS Apple Silicon. This is done
by running a native build in a machine of the target architecture.

At the time the support for producing Apple Silicon builds was added, use of GitHub-hosted Apple Silicon runner machines
was charged by the minute (while use of the other runners is free). In order to avoid excessive expenses, the workflow
was configured so that the Apple Silicon builds were only produced when absolutely necessary. This was done by defining
two sets of job matrix arrays, one for jobs using free runners, and the other for jobs using paid runners. Due to the
limitations of GitHub Actions framework, it was necessary to use workflow environment variables for this purpose.

Since that time, GitHub made free GitHub-hosted Apple Silicon runners available. When the workflow was adjusted to use
that runner, the configuration for the Apple Silicon build job was moved to the free runner job matrix array. The system
for supporting selective use of paid GitHub-hosted runners to produce builds was not removed at that time. This is
reasonable since it is possible the need will arise for using paid runners at some point in the future (e.g., only
legacy older versions of free macOS "Intel" runners are now provided and it is likely that even these will eventually be
phased out forcing us to use the paid runner to produce builds for that target). However, the environment variable for
the paid runner job matrix array data was removed. The absence of that variable made it very difficult to understand the
workflow code for the system.

For this reason, the environment variable is replaced, but empty of data. A comment is added to explain the reason for
this.